+++ src/engine/actions/action/ease-by.ts
@@ -83,5 +83,6 @@ export class EaseBy implements Action {
   public stop(): void {
     this._motion.vel = vec(0, 0);
     this._stopped = true;
+    this._initialized = false;
   }
 }