+++ src/engine/actions/action/scale-by.ts
@@ -74,6 +74,7 @@ export class ScaleByWithOptions implements Action {
   public stop(): void {
     this._motion.scaleFactor = Vector.Zero;
     this._stopped = true;
+    this._started = false;
     this._currentMs = 0;
   }
 
@@ -141,6 +142,7 @@ export class ScaleBy implements Action {
     this._motion.scaleFactor.x = 0;
     this._motion.scaleFactor.y = 0;
     this._stopped = true;
+    this._started = false;
   }
 
   public reset(): void {